Fix new tests
