+++ tests/integration/defs/disaggregated/test_disaggregated.py
@@ -22,7 +22,8 @@ from typing import Callable
 
 import pytest
 import yaml
-from defs.common import wait_for_server
+from defs.common import (
+    revise_disaggregated_server_config_urls_with_free_ports, wait_for_server)
 from defs.conftest import (get_sm_version, llm_models_root, skip_arm,
                            skip_no_hopper)
 from defs.trt_test_alternative import check_call, check_output, popen
@@ -277,7 +278,8 @@ def get_test_config(test_desc, example_dir, test_root):
         raise ValueError(f"Invalid test description: {test_desc}, "
                          f"valid descriptions are: {config_map.keys()}")
 
-    return config_map[test_desc]
+    return revise_disaggregated_server_config_urls_with_free_ports(
+        config_map[test_desc])
 
 
 def get_extra_llm_config(config, suffix, cwd):
